refactor(cr-bridge): extract types to cr-bridge.types.ts, explicit ClusterTenantSpec
Move ClusterTenantOwner (subject now required), ClusterTenantCrSpecPatch, and
new ClusterTenantSpec (extends patch with mandatory owner) to cr-bridge.types.ts.
ClusterTenantCr.spec now typed as ClusterTenantSpec; _BuildOwnerSpec accepts
Partial<ClusterTenantOwner> for the loose session input.
